Hocs 1410 info pipeline (#180)
* Add unique constraint to field_screen table.

* Add unique constraint to case_type_schema table.

* Provide list of cohorts.

* Provide list of official countries.

* HOCS-1061
Changing the deployment script for the build pipeline on the WCS branch.

* HOCS-1173 - Configurable Document Labels (#154)

* Provide list of official countries. (#155)

* Provide list of official countries.

* Enrich list of official countries with list of official territories.

* Code review changes for territories.

* HOCS-1061
Updating the version of 'kd'

* HOCS-1061
Adding the branch checkout for Kube WCS

* HOCS-1061
The application has an exception on startup in the new WCS environment.
Switching on all the debugging to see the issue.

* HOCS-1061
Still no reason given for the connection exception and no helpful logs. Enabling tracing...

* HOCS-1061
This is very annoying. No reason for the connection failure and no trace on the values used for the connection.
Inserting debugs for the application properties for the connections.

* HOCS-1061
Adding debugs for application and environment variables.

* HOCS-1061
Removing the logger debug and trace settings.

* HOCS-1061
Set tracing on for flyway only.

* HOCS-1061
Set tracing on for spring framework.

* Fixing integration test on local builds (#161)

* HOCS-1076-Search for a compensation case

* Fixing integration tests so that local db is not affected when running them

* HOCS-1076-Search for a compensation case (#160)

* HOCS-1061
Removing the tracing, as it is not bringing any value (useless error message)

* HOCS-1061
Adding context listener to show applications properties.

* HOCS-1220 - Work stack changes for WCS (#162)

* HOCS-1220 - Work stack changes for WCS

* HOCS-1220 - Work stack changes for WCS

* HOCS-1296 : change list of cohorts to full English

* WCS-HOCS-1300 add column ordering

* HOCS-1318 : update summary tab

*   HOCS-1320 - remove deadlines from WCS

* HOCS-1370 - Add current stage to case data extract

* WCS-HOCS-1391 added config option

* WCS-HOCS-1391 fix tests

* HOCS-1319 - Case Details Screen

* fixing the deploy issue in qa

* revert - fixing the deploy issue in qa

* HOCS-1387 - Record additional information about claimant's status

* HOCS-1410
Adding the pipeline processing for CS and WCS deployment.

* HOCS-1410
Removed context listener for logging the application properties.

Co-authored-by: Mark Wright <46746088+MarkWrightBjss@users.noreply.github.com>
Co-authored-by: Hubert Raich <52404016+hubert-raich@users.noreply.github.com>
Co-authored-by: Raysonviswas <43213211+Raysonviswas@users.noreply.github.com>
Co-authored-by: Lewis Muirhead <53258753+lewismuirhead@users.noreply.github.com>