Feature/hocs 1751 ci token update (#215)
* HOCS-1751
This is the first phase of the CI changes for the new drone scripts. It is for DEV only.

* HOCS-1751
Updating CI scripts with new tokens for the HOCS service separation.
We no longer have everything under CS. They are separated into namespaces.

* HOCS-1751
Adding the filters for qa and prod deployments.

* HOCS-1751
Implemented the qa and prod deployments. This concludes the changes for all namespaces.