HOCS-2008 - Standard Lines Management - Improvements -added endpoint for fetching simple mapping representing topic uuid to team (display name) responsible for passed in stage - removes some redundant call in test file to 'times(1)' and 'doNothing()'
The logic uses team_link table in the background to find all mappings
for chosen stage where link_type == 'TOPIC'.

Spring data projection has been used to map sql output to SimpleMapItem.
This is so that this generic interface could be reused in the future.
For more details about spring data projection see:
https://www.baeldung.com/spring-data-jpa-projections