HOCS-2095: Add topic name to environment variables
As a result of a production live issue that means that nothing from the hocs-info-service was sending to logging successfully, we have attained that the topic name is not being correctly set within the deployment. This is now added and follows the same path as the other projects we have.