HOCS-2124 Add HTTP proxy settings
The Northern Ireland Assembly publishes its member API over HTTP, not
HTTPS. As we only configured info-service to use a proxy for HTTPS
traffic, this mean that calls to the NI Assembly API always timed out,
as we do not allow non-proxied traffic to exit the network.

This commit adds some Java options to route HTTP calls to the same proxy
as the HTTPS calls. There are some potential security implications here:
it is now possible for other hocs-info-service traffic to exit in
plaintext rather than over TLS. This isn't likely -- everything else
seems to be explicitly set to https://. However, before this change
anything attempting to use HTTP traffic would fail; now it will work.
Just something to be aware of.

Ideally the NI Assembly would upgrade to using HTTPS rather than HTTP.
It's the only use of plaintext within DECS. Although this is a public
API where confidentiality is unimportant, HTTPS also ensures data
integrity in a way that plain HTTP does not.