HOCS-2215 Add explicit internal proxy whitelisting
This commit fixes a misconfiguration that previously was unnoticed as
the system interpreted the misconfiguration loosely, until a later
commit caused the system to start interpreting the argument strictly.

In a previous change we added a whitelist to prevent traffic from within
the namespace being routed over the outbound proxy, but it was specified
using the `http.nonProxyHosts` argument. This seemed to work even though
the proxy was actually only configured for HTTPS (using
`https.proxyHost`).

Commit 4c2648a47366519061a10ab717000af31c3a82dd added a specific proxy
for HTTP traffic (`http.proxyHost`), which seems to have caused the
original `http.notProxyHosts` argument to work solely for the HTTP
requests and not the HTTPS requests as it was previously doing.

This commit adds a `https.nonProxyHosts` argument to JAVA_OPTS to
explicitly state that HTTPS internal traffic should not be routed over
the outbound proxy.

There's an argument to be made that we shouldn't whitelist HTTP internal
traffic, and force every single internal HTTP call to fail -- but I
suspect that would be overly confusing to debug.