HOCS-2229 Use deployment when triggering databuild
Previously we were restarting the top job of hocs-data, which caused
problems. This changes that to fire a `cs-qa` deployment to hocs-data.

The name of this deployment is arbitrary and not currently used by the
downstream pipeline, which only cares about it being *a* deployment and
doesn't differentiate between them.

This commit also shortens the name of the associated build steps, as
they were getting truncated in the Drone CI user interface.