HOCS-2242 Push Docker image on every PR build
This commit makes Drone push interim images on every pull request. This
will mean that developers can deploy them to development environments
manually where required. It's already possible to actually deploy
manually to development environments -- this just ensures there's
something useful to deploy. As Quay.io is free for open-source projects,
there's no real downside to pushing often, but here we only push when a
pull request is opened (which means we don't flood Quay too much, but
also sidestep any sanitisation problems with Docker tags). One potential
downside to this is that someone malicious could open a pull request
with bad things and force a push to our image registry, but that's okay
as long as nobody actually deploys it anywhere.

This commit also changes the behaviour to build a Docker image on most
jobs, which should have been the default anyway -- previously you could
theoretically have built the project successfully, but broken the Docker
build, and there was no way of verifying you hadn't until the post-merge
to master build failed.