HOCS-2138 Fire cron at different time in notprod
Prior to this commit the cronjob was timed to fire at 22:00 in every
environment. This is fine for production, which has pods running then --
but for non-production we scale pods down at 17:30. This means the cron
spins up a pod at 22:00, which fails, so the system spins up a new pod,
which also fails, and so on until it succeeds or hits a limit.