HOCS-2213 Opt cron into backend network policy
The curl in this pod would just time out, as it wasn't permitted by our
network policies to hit the info-service container. This commit places
the temporary pod created by this cron into the hocs-backend group, so
it can access the containers as appropriate.

If we wanted to be more secure we could set a special network policy
that only allows this pod to access info-service and not the other
backend services, but I don't think it's worth the added complexity.