HOCS-2788: send rename notification to notify queue (#280)
* HOCS-2788: add application properties for notify

Add the required properties for notify to be implemented.

* HOCS-2788: add configuration for notify

Add the localstack configuration and aws configuration to allows for 
notify to work locally and in production. Properties set in SQS 
Configuration are part of K8 secrets passed through.

* HOCS-2788: add notify client

Add the notify client that puts messages on the notify queue when a team name is changed.

* HOCS-2788: add notify email call 

After a team name has been renamed, trigger a rename email to be 
sent out to the nominated contacts for that team.