HOCS 2966 merge in main (#286)
* HOCS-2563 : add countries and territories files as local resources.

* HOCS-2563 : corrected logging level and useing server port variable.

* HOCS-2563 add OGL license details and Prettified countries, territories lists

* HOCS 2563 update missing resource handling

* Update README.md

Co-authored-by: Douglas Gardner <douglas.gardner@digital.homeoffice.gov.uk>

* HOCS-2563 add a licence file to the info service project

* HOCS-2457 add description label to Swagger service

* HOCS-2491 remove 'Member' from correspondent type when adding a non-member

* Add selectableInDropdown field to dtos and entities

* HOCS-2457 : add separate swagger config file

* HOCS-2457 add logging to swagger config

* HOCS-2743: improve the updateTeamName method

Presently, the updateTeamName service method does the rename 
blindly. This change does the following: checks if a rename conflict is 
present, evicts the cache of the teams to allow for the update.

* HOCS-2788: send rename notification to notify queue (#280)

* HOCS-2788: add application properties for notify

Add the required properties for notify to be implemented.

* HOCS-2788: add configuration for notify

Add the localstack configuration and aws configuration to allows for 
notify to work locally and in production. Properties set in SQS 
Configuration are part of K8 secrets passed through.

* HOCS-2788: add notify client

Add the notify client that puts messages on the notify queue when a team name is changed.

* HOCS-2788: add notify email call 

After a team name has been renamed, trigger a rename email to be 
sent out to the nominated contacts for that team.

* HOCS-2674: remove team caching (#284)

At present we have individual service caches, which can become out 
of date easily as we run multiple service pods. This causes potential 
inaccuracies of data between pods. This side effect means the user 
does not always see up to to date teams now we support adding 
them dynamically through MUI.

Co-authored-by: Andrew Barker <andrew.barker@digital.homeoffice.gov.uk>
Co-authored-by: abarker66 <71025533+abarker66@users.noreply.github.com>
Co-authored-by: Douglas Gardner <douglas.gardner@digital.homeoffice.gov.uk>
Co-authored-by: agraham19 <andrew.graham@digital.homeoffice.gov.uk>
Co-authored-by: agraham19 <35735033+agraham19@users.noreply.github.com>
Co-authored-by: Liam MacPherson <11508628+LiamMacP@users.noreply.github.com>