HOCS-2986: remove hardcoded emails (#292)
* HOCS-2986: remove hardcoded emails

Currently the whitelisted domains that can be used for user 
additions are hardcoded within `application.properties`. This 
change removes this and allows it to be pulled in through a new 
config map.

* HOCS-2986: allow emails with uppercase

To ensure we support what the real user might do this change does 
the valid check against a semi normalised email address that has 
been changed too lowercase.