HOCS-2239: downscaler annotation addition (#60)
* HOCS-2239: downscaler annotation addition

Presently, when the deployment changes the downscaler annotation 
(that was manually added up to present) gets removed. This causes 
pod to stay up indefinitely, which leads to increased costs and 
potential issues. This change enforced the uptime meta-data 
annotation that enforces this server stays up. This is specified 
differently depending on if we are in prod or notprod.

* HOCS-2239: update member refresh time

CronJobs work based off the time that is set in the `kube-controller-manager`. 
We presume this does not work as intended with BST. This is 
because 1700Z is 18:00 BST and thus would not get called. 
The same is applied to PROD so this time has been changed as well.