Epic/foi mvp rebase (#295)
* HOCS-2982 - Add new domain to user.email.whitelist property (#289)

Domain added = cluster2security.gov.uk

* Fix order of steps

* HOCS-2666 Team members list in M:UI doesn't always update (#290)

Removing the caching fixes one cause, but another issue exists where
adding several users at once creates duplicate groups in keycloak.

Currently batches of users are added by multiple calls from the MUI,
this change batches them into a list to handle them more cleanly and
address the issue above.

* Remove caching on getUserForTeam in UserService
* Process adding of users in batch

* HOCS-2710 Form validation (#288)

Allows validation rules to be passed from DB through the stack

Co-authored-by: Paul Rau <52495258+Paulcr-ho@users.noreply.github.com>

* HOCS-2986: remove hardcoded emails (#292)

* HOCS-2986: remove hardcoded emails

Currently the whitelisted domains that can be used for user
additions are hardcoded within `application.properties`. This
change removes this and allows it to be pulled in through a new
config map.

* HOCS-2986: allow emails with uppercase

To ensure we support what the real user might do this change does
the valid check against a semi normalised email address that has
been changed too lowercase.

* HOCS-2744 Move a team to a different unit via MUI (#294)

* HOCS-2744 Move a team to a different unit via MUI

* Add patch endpoint to update multiple team fields

* rebase

* HOCS-2885 Automatically deploy to gamma on push

This adds a step to Drone that fires on push to the FOI epic branch,
which triggers a deployment to the hocs-gamma environment.

Previously we were deploying manually to hocs-gamma, and using a fixed
branch-based tag, which was causing inconsistencies and confusion. Here
we deploy a tag based on the Drone build number, which is easier to
track. If we wanted, we could deploy using the SHA hash for a similar
effect.

Co-authored-by: Phil Barton <phil.barton@digital.homeoffice.gov.uk>
Co-authored-by: Douglas Gardner <douglas.gardner@digital.homeoffice.gov.uk>
Co-authored-by: Robert Deniszczyc <72561986+robertdeniszczyc2@users.noreply.github.com>
Co-authored-by: Paul Rau <52495258+Paulcr-ho@users.noreply.github.com>
Co-authored-by: Liam MacPherson <11508628+LiamMacP@users.noreply.github.com>