HOCS-3631: case deadline dates (#323)
* HOCS-3631: change date utils

As the only current use case for the DateUtils class is to check if a
date is a non-working day, this change pulls that logic into this file.
As a result the isWeekend method can be made private.

* HOCS-3631: change deadline implementation

If a date was added on a weekend then the counting would start
from that date. This change moves this date forward to the next 
non-working date before the calculation is started.

* HOCS-3631: use new DateUtil

The CaseTypeService currently uses the old isWeekend method, 
as a simplified method has been added that checks if a date is a 
non-working day, this is now used.