Remove data version requirement
In some instances, releasing seperate data versions to info is required.
For mainline environments this isn't possible and ends up requiring
manual deployment through modifying K8 config. This change removes this
limitation as by default it uses version and if someone was to do this,
they should have a valid reason.