HOCS-4421: dev deployment should use commit hash (#430)
At the minute when a change is merged into main it uses the `build_XXXX`
tag to highlight the version. We instead want to drive this using the 
Git commit sha. This allows us too checkout the version from within the
pipeline and use the point in time deployment config. 

Co-authored-by: Liam MacPherson <liammacpherson1996@gmail.com>