HOCS-4696: use GitHub SemVer actions
This change removes the use of the old hocs-version-bot in favour of
new GitHub actions.

When an action happens on a PR an action is run that checks that the PR
is labeled with one of the following: Patch, Minor, or Major.
If one of these isn't present the PR will be blocked till one is added.
This enables us to immediately identify the scale of the change.

On an successful merge, an second action is ran that uses specified
label to update the SemVer version for the service - automatically
creating a Git Tag with that version.

This triggers a drone build that automatically deploys this new version
to the dev environments. This triggers a downstream build to CS-DEV and
WCS-DEV to tag that image in ECR. The deployment to the environment
awaits the success of the build and push and the tagging of data.