HOCS-5057 Tell info-service about the HTTP proxy
The NI Assembly data service info-service looks up is only available via
plain old HTTP (:80) and 404s if you try HTTPS (:443). Prior to this
commit hocs-info-service was configured to route outbound HTTPS requests
through the outbound proxy, but not HTTP requests, meaning that the NI
Assembly request times out.

This commit tells hocs-info-service explicitly to go via the outbound
proxy for HTTP requests as well.

This commit also fixes a typo in the config: the non-proxy host list
only applied to HTTP but was probably meant to apply only to HTTPS.